+++ src/plone/jsonapi/__init__.py
@@ -1,10 +1,1 @@
-# -*- coding: utf-8 -*-
-#
-# File: __init__.py
-
-__author__ = 'Ramon Bartl <ramon.bartl@googlemail.com>'
-__docformat__ = 'plaintext'
-
-from browser import router
-
-# vim: set ft=python ts=4 sw=4 expandtab :
+__import__('pkg_resources').declare_namespace(__name__)