+++ src/plone/jsonapi/core/browser/router.py
@@ -14,7 +14,10 @@ from interfaces import IRouteProvider
 
 logger = logging.getLogger("plone.jsonapi.router")
 
+
 class Router(object):
+    """ API Router
+    """
 
     def __init__(self):
         logger.info("DefaultRouter::__init__")
@@ -27,10 +30,14 @@ class Router(object):
     def initialize(self, context, request):
         """ called by the API Framework
         """
+        logger.info("DefaultRouter.initialize: context=%r request=%r" % (context, request))
+
         self.context = context
         self.request = request
+
         self.environ = request.environ
         self.http_host = request["HTTP_HOST"]
+        self.url = request.getURL()
 
         if self.is_initialized:
             return
@@ -103,10 +110,22 @@ class Router(object):
         (values=None, method=None, force_external=False, append_unknown=True)
         see: http://werkzeug.pocoo.org/docs/routing/#werkzeug.routing.MapAdapter.build
         """
-        # XXX: this is hacky - we need to retain the API base url to create correct links.
-        path_info = self.environ["PATH_INFO"]
-        base, _, _ = path_info.partition("API")
-        adapter = self.get_adapter(script_name=base + "API")
+
+        # XXX: this is all a little bit hacky, especially when it comes to virtual hosting.
+
+        spp = self.request.physicalPathFromURL(self.url)
+
+        # find the API view root
+        path = []
+        for el in spp:
+            path.append(el)
+            if el == "API" or el == "@@API":
+                break
+
+        virt_path = self.request.physicalPathToVirtualPath(path)
+        script_name = self.request.physicalPathToURL(virt_path, relative=1)
+
+        adapter = self.get_adapter(script_name=script_name)
         return adapter.build(endpoint, **options)
 
     def __call__(self, context, request, path):